+++ trunk/game/sprites/monsters/PriestSprite.cs
@@ -34,7 +34,7 @@ namespace AbrahmanAdventure.sprites
         /// <param name="xPosition">x position</param>
         /// <param name="yPosition">y position</param>
         /// <param name="random">random number generator</param>
-        public PriestSprite(double xPosition, double yPosition, Random random)
+        public PriestSprite(float xPosition, float yPosition, Random random)
             : base(xPosition, yPosition, random)
         {
             if (deadSurface == null)
@@ -136,12 +136,12 @@ namespace AbrahmanAdventure.sprites
             return false;
         }
 
-        protected override double BuildJumpProbability()
+        protected override float BuildJumpProbability()
         {
             return 0.2;
         }
 
-        protected override double BuildChangeDirectionNoAiCycleLength()
+        protected override float BuildChangeDirectionNoAiCycleLength()
         {
             return 0.0;
         }
@@ -151,7 +151,7 @@ namespace AbrahmanAdventure.sprites
             return false;
         }
 
-        protected override double BuildSafeDistanceAi()
+        protected override float BuildSafeDistanceAi()
         {
             return 0.0;
         }
@@ -166,67 +166,67 @@ namespace AbrahmanAdventure.sprites
             return false;
         }
 
-        protected override double BuildMaxHealth()
+        protected override float BuildMaxHealth()
         {
             return 0.5;
         }
 
-        protected override double BuildJumpingTime()
+        protected override float BuildJumpingTime()
         {
             return 10.0;
         }
 
-        protected override double BuildWalkingCycleLength()
+        protected override float BuildWalkingCycleLength()
         {
             return 3;
         }
 
-        protected override double BuildWalkingAcceleration()
+        protected override float BuildWalkingAcceleration()
         {
             return 0.01;
         }
 
-        protected override double BuildMaxWalkingSpeed()
+        protected override float BuildMaxWalkingSpeed()
         {
             return 0.20;
         }
 
-        protected override double BuildMaxRunningSpeed()
+        protected override float BuildMaxRunningSpeed()
         {
             return 0.20;
         }
 
-        protected override double BuildStartingJumpAcceleration()
+        protected override float BuildStartingJumpAcceleration()
         {
             return 8.0;
         }
 
-        protected override double BuildAttackingTime()
+        protected override float BuildAttackingTime()
         {
             return 4;
         }
 
-        protected override double BuildHitTime()
+        protected override float BuildHitTime()
         {
             return 32;
         }
 
-        protected override double BuildAttackStrengthCollision()
+        protected override float BuildAttackStrengthCollision()
         {
             return 0.5;
         }
 
-        protected override double BuildWidth(Random random)
+        protected override float BuildWidth(Random random)
         {
             return 1.0;
         }
 
-        protected override double BuildHeight(Random random)
+        protected override float BuildHeight(Random random)
         {
             return 1.9;
         }
 
-        public override Surface GetCurrentSurface(out double xOffset, out double yOffset)
+        public override Surface GetCurrentSurface(out float xOffset, out float yOffset)
         {
             xOffset = yOffset = 0;
             if (!IsAlive)
